% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getMode}
\alias{getMode}
\title{Function to return the mode of a vector of values}
\usage{
getMode(v)
}
\arguments{
\item{v}{vector of numeric values}
}
\value{
a numeric value
}
\description{
This is an internal function. Please use cautiously if calling directly. Returns the mode from numeric vector. Example usage: \code{getMode(c(0,1,1))}
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
