% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noFalseCalls.R
\name{noFalseCalls}
\alias{noFalseCalls}
\title{Function to calculate the number of false calls for a design for a go-no go threshold}
\usage{
noFalseCalls(dataset, GoNoGoThreshold)
}
\arguments{
\item{dataset}{dataset obtained as output from the 'varying_PDXn_PDXr' function}

\item{GoNoGoThreshold}{go-no go threshold}
}
\value{
vector with three elements:

- numeric value indicating the number of experiments simulated

- numeric value indicating the number of experiments which were above the go-no go threshold

- numeric value indicating the FPR
}
\description{
This is an internal function. Please use cautiously if calling directly. Returns the number of false calls from a simulation study exploring the impact of varying PDXn and PDXr for an underlying Biol_RR for a particularly go-no go threshold.
A false call can only arise in the situation where the underlying Biol_RR is below the go-no go threshold.
Example usage: \code{noFalseCalls(ImpactVarying_PDXn_PDXr_BRR, GoNoGoThreshold=30)}
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
