% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noMissedCalls.R
\name{noMissedCalls}
\alias{noMissedCalls}
\title{Function to calculate the number of missed calls for a design for a go-no go threshold}
\usage{
noMissedCalls(dataset, GoNoGoThreshold)
}
\arguments{
\item{dataset}{dataset obtained as output from the 'varying_PDXn_PDXr' function}

\item{GoNoGoThreshold}{go-no go threshold}
}
\value{
vector with three elements:

- numeric value indicating the number of experiments simulated

- numeric value indicating the number of experiments which were below the go-no go threshold

- numeric value indicating the percent of missed calls
}
\description{
This is an internal function. Please use cautiously if calling directly.
Returns the number of missed calls from a simulation study exploring the impact of varying PDXn and PDXr for an underlying Biol_RR for a particularly go-no go threshold.
A missed call can only arise in the situation where the underlying Biol_RR exceeds the go-no go threshold. Example usage: \code{noMissedCalls(ImpactVarying_PDXn_PDXr_BRR, GoNoGoThreshold=30)}
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
