% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeInSingleExperiment.R
\name{outcomeInSingleExperiment}
\alias{outcomeInSingleExperiment}
\title{Function to summarise the results for a single simulation representing one experiment}
\usage{
outcomeInSingleExperiment(df, PDXn, PDXr, C_Acc, Biol_RR)
}
\arguments{
\item{df}{data frame from callsInSingleExperiment}

\item{PDXn}{PDXn}

\item{PDXr}{PDXr}

\item{C_Acc}{the classification accuracy (numeric value between 0 and 1)}

\item{Biol_RR}{Biol_RR}
}
\value{
a vector with 8 values that captures the input design and the estimated response rate for that design from a single simulation
}
\description{
This is an internal function. Please use cautiously if calling directly. From a simulation of a single experiment, the estimated response rate is determined and captured with the meta data (e.g. PDXn, PDXr) for that experiment.
Example usage: \code{outcomeInSingleExperiment(df=outcomeInSingleExperiment_1, PDXn=8, PDXr=3, C_Acc=0.95, Biol_RR=30)}
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
