% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeMultipleExperiments.R
\name{outcomeMultipleExperiments}
\alias{outcomeMultipleExperiments}
\title{Function to run simulations to mimic population PDX studies for a defined scenario}
\usage{
outcomeMultipleExperiments(PDXn, PDXr, C_Acc, Biol_RR, iterations)
}
\arguments{
\item{PDXn}{PDXn}

\item{PDXr}{PDXr}

\item{C_Acc}{the classification accuracy (numeric value between 0 and 1)}

\item{Biol_RR}{Biol_RR}

\item{iterations}{no of experiments to simulated}
}
\value{
a dataframe where each row represents the results from a simulation mimicking an individual experiment for a particular design with meta data returned to describe the experimental design
}
\description{
This is an internal function. Please use cautiously if calling directly. Simulations are used to mimic population PDX studies for specified values of PDXn, PDXr, Biol_RR and C_Acc.
Example usage: \code{outcomeMultipleExperiments(PDXn=8, PDXr=3, C_Acc=0.95, Biol_RR=30, iterations=500)}
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
