% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFalsepositive.R
\name{plotFalsepositive}
\alias{plotFalsepositive}
\title{A function to visualise the false postive rate as a function of PDXn and PDXr}
\usage{
plotFalsepositive(data)
}
\arguments{
\item{data}{data frame with four columns which indicate the PDXn, PDXr, Biol_RR and the FPR for a specified go-no go threshold}
}
\value{
a graphic visualisation
}
\description{
This is an internal function. Please use cautiously if calling directly.
A visualisation of the false positive rate behaviour from the simulations
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
