% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{A function to visualise the sensitivity as a function of PDXn and PDXr}
\usage{
plotSensitivity(data)
}
\arguments{
\item{data}{data frame with four columns which indicate the PDXn, PDXr, Biol_RR and the MissedCalls for a specified go-no go threshold}
}
\value{
a graphic visualisation
}
\description{
This is an internal function. Please use cautiously if calling directly.
A visualisation of the sensitivity from the simulations
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
