% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{populationPDXdesignApp}
\alias{populationPDXdesignApp}
\title{Function \code{populationPDXdesignApp}}
\usage{
populationPDXdesignApp()
}
\description{
Runs the 'shiny' app.
}
\examples{
if (interactive()) {
populationPDXdesignApp()
}

}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
