% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{server}
\alias{server}
\title{'shiny' app server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
This is an internal function. Please use cautiously if calling directly
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
