% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ui}
\alias{ui}
\title{'shiny' app user interface function}
\usage{
ui()
}
\description{
This is an internal function. Please use cautiously if calling directly.
}
\author{
Maria Luisa Guerriero, \email{maria.guerriero@astrazeneca.com}

Natasha A. Karp, \email{natasha.karp@astrazeneca.com}
}
